Feature/service login and NewGoal api

+++ data/src/main/java/com/eroom/data/entity/Participant.kt
@@ -3,7 +3,7 @@ package com.eroom.data.entity
 import com.fasterxml.jackson.annotation.JsonProperty
 
 data class Participant(
-    @JsonProperty("id") var id: Int,
+    @JsonProperty("id") var id: Long,
     @JsonProperty("name") var name: String,
     @JsonProperty("imageUrl") var imageUrl: String
 )
